requirements.txt
urllib.request
Pillow